+++ CHANGELOG.md
@@ -6,3 +6,4 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - [2025-12-21 01:48:42] Added search functionality to the Generation History page to filter posts by title.
+- [2024-05-22 10:00:00] Made generated topic titles editable in the Planner before scheduling.